So I worked with some stuff.
To start, I made it so that the arrival eventTypes
are processed and added to the appropriate queue.

Next, I added a light change event that occurs
every 30 i time. Another event is added
whenever the light-change occurs.

Now that the light changes, we just have to make it
so that we move x cars behind a green light to the
apropriate queue.

Also, arrivals now come from all directions.

- Jake